Split type_check into type_formation and term_formation
