+++ ast/src/validation/mod.rs
@@ -1,1 +1,10 @@
-pub mod type_checker;
+pub mod base;
+pub mod type_formation;
+pub mod term_formation;
+
+#[cfg(test)]
+mod tests {
+    #[test]
+    fn example() {
+    }
+}
